+++ app/route/v1/map/map-api-id.js
@@ -1,0 +1,17 @@
+let express = require('express');
+let router = express.Router();
+let getMapById = require('./get-map-by-id.js');
+
+/**
+ * api/v1/mapのルーティングを設定
+ * @param {object} router - express.Routerのobject
+ * @return {object} router - 引数と同じ，ルーティングを設定して返す
+ */
+let setRoutingForMapIdApi = (router) => {
+    return router.get('/:id', (req, res) => {
+        let mapId = req.params.id;
+        res = getMapById(mapId, res);
+    })
+}
+
+module.exports = setRoutingForMapIdApi(router);
